fix: Move `types/test_util.go:MakeBlock` into `types/block.go`
